// src/renderer/src/app/enums/shortcut-action.ts
/**
 * @license
 * Copyright (c) Noldova.
 *
 * This source code is licensed under the license found in the
 * LICENSE file in the root directory of this source tree.
 */

export enum ShortcutAction {
  NewConversation = "NewConversation",
  OpenFolder = "OpenFolder",
  OpenSettings = "OpenSettings",
  ShowShortcuts = "ShowShortcuts",
  ToggleSidebar = "ToggleSidebar",
  Back = "Back",
  Forward = "Forward",
  Search = "Search",
  ToggleExplorer = "ToggleExplorer",
  ToggleChanges = "ToggleChanges",
  ToggleActivity = "ToggleActivity",
  ToggleBottomDock = "ToggleBottomDock",
  CloseDocument = "CloseDocument",
  FocusComposer = "FocusComposer",
  StopOrBack = "StopOrBack",
  PreviousConversation = "PreviousConversation",
  NextConversation = "NextConversation",
  Send = "Send",
  NewLine = "NewLine"
}